Move LoggerDouble to separate file for reuse
