+++ test/external_asset_pipeline/logger_double.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module ExternalAssetPipeline
+  class LoggerDouble
+    attr_reader :messages
+
+    def warn(message)
+      @messages ||= []
+      @messages << message
+    end
+  end
+end
